Added all entities

+++ Model/Consumption.java
@@ -2,13 +2,13 @@ package Model;
 
 public class Consumption
 {
-    private Long consumptionId;
-    private Long meterId;
+    private int consumptionId;
+    private int meterId;
     private Double consumptionValue;
     private String readingDate;
     private String billingPeriod;
 
-    public Consumption(Long consumptionId, Long meterId, Double consumptionValue, String readingDate, String billingPeriod)
+    public Consumption(int consumptionId, int meterId, Double consumptionValue, String readingDate, String billingPeriod)
     {
         this.consumptionId = consumptionId;
         this.meterId = meterId;
@@ -18,10 +18,10 @@ public class Consumption
     }
 
     // Getters and Setters
-    public Long getConsumptionId() { return consumptionId; }
-    public void setConsumptionId(Long consumptionId) { this.consumptionId = consumptionId; }
-    public Long getMeterId() { return meterId; }
-    public void setMeterId(Long meterId) { this.meterId = meterId; }
+    public int getConsumptionId() { return consumptionId; }
+    public void setConsumptionId(int consumptionId) { this.consumptionId = consumptionId; }
+    public int getMeterId() { return meterId; }
+    public void setMeterId(int meterId) { this.meterId = meterId; }
     public Double getConsumptionValue() { return consumptionValue; }
     public void setConsumptionValue(Double consumptionValue) { this.consumptionValue = consumptionValue; }
     public String getReadingDate() { return readingDate; }
