+++ Model/Department.java
@@ -2,19 +2,19 @@ package Model;
 
 public class Department
 {
-    private Long departmentId;
+    private int departmentId;
     private String departmentName;
 
-    public Department(Long departmentId, String departmentName)
+    public Department(int departmentId, String departmentName)
     {
         this.departmentId = departmentId;
         this.departmentName = departmentName;
     }
 
     // Getters and Setters
-    public Long getDepartmentId() { return departmentId; }
+    public int getDepartmentId() { return departmentId; }
 
-    public void setDepartmentId(Long departmentId) { this.departmentId = departmentId; }
+    public void setDepartmentId(int departmentId) { this.departmentId = departmentId; }
 
     public String getDepartmentName() { return departmentName; }
